+++ test/expression/node/FunctionNode.test.js
@@ -4,99 +4,124 @@ var assert = require('assert'),
     math = require('../../../index')(),
     Node = require('../../../lib/expression/node/Node'),
     ConstantNode = require('../../../lib/expression/node/ConstantNode'),
-    OperatorNode = require('../../../lib/expression/node/OperatorNode'),
-    FunctionNode = require('../../../lib/expression/node/FunctionNode'),
+    SymbolNode = require('../../../lib/expression/node/SymbolNode'),
     RangeNode = require('../../../lib/expression/node/RangeNode'),
-    SymbolNode = require('../../../lib/expression/node/SymbolNode');
+    FunctionNode = require('../../../lib/expression/node/FunctionNode');
 
 describe('FunctionNode', function() {
 
   it ('should create a FunctionNode', function () {
-    var n = new FunctionNode('f', ['x'], new Node());
+    var s = new SymbolNode('sqrt');
+    var c = new ConstantNode(4);
+    var n = new FunctionNode(s, [c]);
     assert(n instanceof FunctionNode);
     assert(n instanceof Node);
     assert.equal(n.type, 'FunctionNode');
   });
 
   it ('should throw an error when calling without new operator', function () {
-    assert.throws(function () {FunctionNode('f', ['x'], new Node())}, SyntaxError);
+    var s = new SymbolNode('sqrt');
+    var c = new ConstantNode(4);
+    assert.throws(function () {FunctionNode(s, [c])}, SyntaxError);
   });
 
-  it ('should throw an error on wrong constructor arguments', function () {
-    assert.throws(function () {new FunctionNode()}, TypeError);
-    assert.throws(function () {new FunctionNode('a')}, TypeError);
-    assert.throws(function () {new FunctionNode('a', ['x'])}, TypeError);
-    assert.throws(function () {new FunctionNode('a', [2], new Node())}, TypeError);
-    assert.throws(function () {new FunctionNode(null, ['x'], new Node())}, TypeError);
+  it ('should throw an error when calling with wrong arguments', function () {
+    var s = new SymbolNode('sqrt');
+    var c = new ConstantNode(4);
+    assert.throws(function () {new FunctionNode('sqrt', [])}, TypeError);
+    assert.throws(function () {new FunctionNode(s, [2, 3])}, TypeError);
+    assert.throws(function () {new FunctionNode(s, [c, 3])}, TypeError);
   });
 
   it ('should compile a FunctionNode', function () {
-    var a = new ConstantNode(2);
-    var x = new SymbolNode('x');
-    var o = new OperatorNode('+', 'add', [a, x]);
-    var n = new FunctionNode('f', ['x'], o);
+    var s = new SymbolNode('sqrt');
+    var c = new ConstantNode(4);
+    var n = new FunctionNode(s, [c]);
 
-    var expr = n.compile(math);
     var scope = {};
-    var f = expr.eval(scope);
-    assert.equal(typeof scope.f, 'function');
-    assert.equal(scope.f(3), 5);
-    assert.equal(scope.f(5), 7);
-    assert.throws(function () {scope.f()}, SyntaxError);
-    assert.throws(function () {scope.f(2, 3)}, SyntaxError);
+    assert.equal(n.compile(math).eval(scope), 2);
+  });
 
+  it.skip ('should compile a FunctionNode acting on a matrix', function () {
+    var s = new SymbolNode('a');
+    var n = new FunctionNode(s, [
+      new ConstantNode(2),
+      new ConstantNode(1)
+    ]);
+
+    var scope = {
+      a: [[1, 2], [3, 4]]
+    };
+    assert.equal(n.compile(math).eval(scope), 3);
+  });
+
+  it.skip ('should compile a FunctionNode acting on a matrix (2)', function () {
+    var s = new SymbolNode('a');
+    var n = new FunctionNode(s, [
+      new ConstantNode(2),
+      new RangeNode([
+        new ConstantNode(1),
+        new SymbolNode('end')
+      ])
+    ]);
+
+    var scope = {
+      a: [[1, 2], [3, 4]]
+    };
+    assert.deepEqual(n.compile(math).eval(scope), [3, 4]);
+  });
+
+  it.skip ('should compile a FunctionNode acting on a matrix (3)', function () {
+    var s = new SymbolNode('a');
+    var n = new FunctionNode(s, [
+      new ConstantNode(2),
+      new RangeNode([
+        new SymbolNode('end'),
+        new ConstantNode(1),
+        new ConstantNode(-1)
+      ])
+    ]);
+
+    var scope = {
+      a: [[1, 2], [3, 4]]
+    };
+    assert.deepEqual(n.compile(math).eval(scope), [4, 3]);
   });
 
   it ('should find a FunctionNode', function () {
-    var a = new ConstantNode(2);
-    var x = new SymbolNode('x');
-    var o = new OperatorNode('+', 'add', [a, x]);
-    var n = new FunctionNode('f', ['x'], o);
+    var a = new SymbolNode('a'),
+        b = new ConstantNode(2),
+        c = new ConstantNode(1);
+    var n = new FunctionNode(a, [b, c]);
 
     assert.deepEqual(n.find({type: FunctionNode}),  [n]);
-    assert.deepEqual(n.find({type: SymbolNode}),    [x]);
+    assert.deepEqual(n.find({type: SymbolNode}),    [a]);
     assert.deepEqual(n.find({type: RangeNode}),     []);
-    assert.deepEqual(n.find({type: ConstantNode}),  [a]);
-    assert.deepEqual(n.find({type: ConstantNode, properties: {value: '2'}}),  [a]);
+    assert.deepEqual(n.find({type: ConstantNode}),  [b, c]);
+    assert.deepEqual(n.find({type: ConstantNode, properties: {value: '2'}}),  [b]);
     assert.deepEqual(n.find({type: ConstantNode, properties: {value: '4'}}),  []);
   });
 
-  it ('should throw an error when creating a FunctionNode with a reserved keyword', function () {
-    assert.throws(function () {
-      new FunctionNode('end', ['x'], new Node());
-    }, /Illegal function name/)
-  });
-
-  it ('should find a FunctionNode without expression', function () {
-    var e = new FunctionNode('f', ['x'], new Node());
-
-    assert.deepEqual(e.find({type: FunctionNode}),  [e]);
-    assert.deepEqual(e.find({type: SymbolNode}),    []);
-  });
-
   it ('should match a FunctionNode', function () {
-    var a = new FunctionNode('f', ['x'], new Node());
+    var a = new FunctionNode(new Node(), []);
     assert.equal(a.match({type: FunctionNode}),  true);
     assert.equal(a.match({type: SymbolNode}), false);
   });
 
   it ('should stringify a FunctionNode', function () {
-    var a = new ConstantNode(2);
-    var x = new SymbolNode('x');
-    var o = new OperatorNode('+', 'add', [a, x]);
-    var n = new FunctionNode('f', ['x'], o);
+    var s = new SymbolNode('sqrt');
+    var c = new ConstantNode(4);
+    var n = new FunctionNode(s, [c]);
 
-    assert.equal(n.toString(), 'function f(x) = 2 + x');
+    assert.equal(n.toString(), 'sqrt(4)');
   });
 
-  it ('should LaTeX a FunctionNode', function() {
-    var a = new ConstantNode(2);
-    var x = new SymbolNode('x');
-    var o = new OperatorNode('/', 'divide', [x, a]);
-    var p = new OperatorNode('^', 'pow', [o, a]);
-    var n = new FunctionNode('f', ['x'], p);
+  it ('should LaTeX a FunctionNode', function () {
+    var s = new SymbolNode('sqrt');
+    var c = new ConstantNode(4);
+    var n = new FunctionNode(s, [c]);
 
-    assert.equal(n.toTex(), 'f\\left({x}\\right)={\\left({\\frac{x}{2}}\\right)^{2}}');
+    assert.equal(n.toTex(), '\\sqrt{4}');
   });
 
 });